// src/models/CourseViewModel.ts
export type CourseViewModel = {
    /**
     * Представление курса с обновленной информацией.
     * id - Идентификатор курса.
     */
    id: number;
    /**
     * Представление курса с обновленной информацией.
     * title - Новое название для курса.
     */
    title: string;
};